+++ setup.cfg
@@ -11,6 +11,8 @@ install_requires =
     pandas
     openpyxl
     matplotlib
+    pytest
+    mypy
 
 [options.packages.find]
 where=src
