+++ src/serena/lmv_new.py
@@ -1,0 +1,18 @@
+"""
+Entry point for local minma variation implementation of ensemble variatioon
+copyright 2023 GrizzlyEngineer
+"""
+
+class LocalMinimaVariation():
+    """
+    Class that is access point for LMV metric I developd that utilizes the 
+    ensemble variation algorithm I developed
+    """
+    def __init__(self) -> None:
+        pass
+
+    def run_lmv(self):
+        pass
+
+    def single_temp_lmv(self, temp: int):
+        pass
